Moved CompilerValidator to Util, add error count
